Added TestBase class, GroupDeletionTest. Refactoring of GroupCreationTest class has been donÑe.
